+++ Console.php
@@ -39,6 +39,7 @@
     </div>
 </div>
 <script src="https://code.jquery.com/jquery-3.5.1.js"></script>
+<script type="text/javascript" src="js/Gethtml-Ajax.js"></script>
 <script type="text/javascript" src="js/Console.js"></script>
 </body>
 </html>